fix price type

+++ api/models/Listings.js
@@ -27,7 +27,7 @@ module.exports = {
       type:'string'
     },
     price:{
-      type:'integer'
+      type:'float'
     },
     province:{
       type:'string'
